+++ MobileFuse/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.9.3.0
+* Certified with MobileFuse SDK 1.9.3.
+* Distributed as a static framework within the XCFramework.
+* Updated ad display failed error code.
+  
 ## 1.9.2.1
 * Removed requirement of the title asset for native banners and MRECs.